tools/system_tests/requirements.txt
# System Test Requirements
# Install with: pip install -r requirements.txt

pytest>=7.0.0
requests>=2.28.0
pyserial>=3.5




